Added GetTLSCertificateFromConnection() which derives TLS cert from TLS session.
There is already a GetTLSCertificate() which uses the API. But for some reason this
requires you to connect with Admin credentials.  Which is a bit odd given that if you
are already talking to the server over HTTPS then it has already sent you its TLS
certificate.  So it isn't like it's a big secret what the TLS certificate is :-).